+++ src_test/fold/test_deriving_fold.cppo.ml
@@ -7,8 +7,7 @@ let test_btree ctxt =
   let btree  = (Node (Node (Leaf, 3, Leaf), 1, Node (Leaf, 2, Leaf))) in
   assert_equal ~printer:string_of_int 6 (fold_btree (+) 0 btree)
 
-type 'a reflist = 'a Pervasives.ref list
-[@@ocaml.warning "-3"]
+type 'a reflist = 'a Stdlib.ref list
 [@@deriving fold]
 
 let test_reflist ctxt =