+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+(unreleased)
+------------
+
+* Introduce `Ppx_deriving_runtime.Stdlib` with OCaml >= 4.07. This module
+  already exists in OCaml < 4.07 but was missing otherwise.
+
 5.2.1 (02/02/2021)
 ------------------